ebuild to have systemd working.
dbus and udev are just required to add the systemd use flag that gives
the following option to their configure scripts:

   --with-systemdsystemunitdir=/lib/systemd/system

these are supported at least in udev-162 and dbus-1.3.2, which are
required for proper systemd integration (udev is even automatically
started by systemd).

The systemd-9999.ebuild installs, but my system does not run xdm and I
still need to investigate why. Maybe one of the reason is some missing
service that need to be enabled.

My setup does not use initramfs/initrd, so I need to add
remount-rootfs.service.